Release V1.09. Download under releases.
Fixed Char generation Buttons for UW2 to be readable.
Set line widths for UW2 message scrolls and conversation window
Fixed setting of the play_poison variable.
Disabled editor and quickstart when ran outside of editor.
Fixed loading of materials to properly detect if a modded texture is loaded and if so load an appropiate shader.
Fixed the poison effect of toadstools.
Picking an item up in mouselook will automatically switch control mode out of mouselook 
Fixed loading of paths to support paths without trailing seperator.